Update README.md maven co-ordinates to the current release biz.neustar : browsermob-proxy : 2.0-beta-5
Also specifies test scope as most people won't want their production code to depend on browsermob. If you're building an app as an extension to browsermob then you probably know enough to remove the scope. If you are copying and pasting stuff off the internet to try and get some performance testing working, then hats off to you, but you've probably got other stuff on your mind and may not notice.

In other other news, my project has a dependency on the 2.19.0 selenium release, so I had to add an exclusion on the selenium-api dependency to make my IDE happy.

Thanks for the codez.